start  working towards a clientId for items
a clientId can be generated and used for an before and after its
upload.  This gives us a consistent way to handle pasting and image
drops with tagging.  There is the notion that something could be tagged
before the item is even uploaded.. will think about how to decouple
that…